Moved memsink tests out into separate file
